Set locale during doc build to C.utf8
Depending on the mock version and chroot used, some locales may or
may be not be available during package build. In the case of the
Fedora Rawhide chroot when building on CentOS 7, only a few locales
are available, and en_EN.utf8 is not one of them. This causes doc
build to fail, since gnocchiclient uses some Unicode strings [1].

C.utf8 seems to be universally available, so let's use it.

[1] - https://github.com/gnocchixyz/python-gnocchiclient/blob/4.0.0/gnocchiclient/utils.py#L91-L102

Change-Id: I7350a6dea65af8f11831aae897cd5fb97d3e207b